Issue update at object's position when AttachCameraToDynamicObject or EditDynamicObject is used
Partially fixes #49. The object may be immediately streamed out on the
next update, but at least now the natives won't just silently fail when
the objects are out of range.

Footnote: The only way to ensure they won't immediately stream out is to
make the objects static (i.e., set a negative streaming distance). This
could be done automatically as well, but there's no real way to know
when the object's original streaming distance should be restored.
Therefore, I will make this the responsibility of the scripter. I'll
also add a native that allows the object's static state to be easily
toggled.