Allow items to be streamed only in specified areas
If one or more areas are specified for an item, the item will only be
streamed if the player is in one of those areas. This adds an 'areaid'
parameter to all items ('areas' for the extended natives) and
E_STREAMER_AREA_ID (located at the very top of the enum in streamer.inc)
to the data manipulation natives.

Resolves #71.