Add item priority support
This adds a 'priority' parameter to all of the creation natives (except
for areas) as well as E_STREAMER_PRIORITY to the data manipulation
natives. The priority can be any number. It is set to 0 by default.
Items with higher priorities will always be streamed before items with
lower priorities.

Resolves #86.