Fix actor bugs
Actors can only be in one virtual world at a time, apparently, unlike
pickups. If -1 is specified, the actor will just be put in the default
virtual world, which is 0.

There was a problem with sampgdk that was making the server crash when
DestroyActor was called in AmxUnload. (I think it was trying to hook
OnActorStreamOut after the script had already been unloaded, but I'm not
entirely sure.) I fixed this by destroying the actors later in
ProcessTick.

This should solve #174.