Move discoverPickups and discoverActors in to the global tickrate
One of the problem after the chunk streaming integration is what the actors and pickups sometimes disappearing. After some research I determined what the problem causing by player tickrate, `haveAllPlayersCheckedActors` and `haveAllPlayersCheckedPickups` checks. With high player tickrate and/or high ammount of newely connected players these functions will `return false` until some "window" without new player connections. So, `streamPickups` and `streamActors` functions may not be called for a long time.

This patch moves `discoverPickups` and `discoverActors` to the global tickrate counter, which means what the pickups and actors are no longer depend on player's tickrate. And this fixes the problem.